# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tecnimotos API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/ordenes-trabajo": {
      "post": {
        "tags": [
          "taller"
        ],
        "summary": "EP-TAL-01: Abrir orden de trabajo",
        "operationId": "abrir_ot_v1_ordenes_trabajo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbrirOTRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Abrir Ot V1 Ordenes Trabajo Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ordenes-trabajo/{ot_id}/repuestos": {
      "post": {
        "tags": [
          "taller"
        ],
        "summary": "EP-TAL-02: Agregar repuesto a orden de trabajo",
        "operationId": "agregar_repuesto_v1_ordenes_trabajo__ot_id__repuestos_post",
        "parameters": [
          {
            "name": "ot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgregarRepuestoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Agregar Repuesto V1 Ordenes Trabajo  Ot Id  Repuestos Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ordenes-trabajo/{ot_id}/aprobar-lista": {
      "post": {
        "tags": [
          "taller"
        ],
        "summary": "EP-TAL-03: Aprobar lista → EN_EJECUCION",
        "operationId": "aprobar_lista_v1_ordenes_trabajo__ot_id__aprobar_lista_post",
        "parameters": [
          {
            "name": "ot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Aprobar Lista V1 Ordenes Trabajo  Ot Id  Aprobar Lista Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ordenes-trabajo/{ot_id}/confirmar-adicional": {
      "post": {
        "tags": [
          "taller"
        ],
        "summary": "EP-TAL-04: Cliente confirma costo adicional manual",
        "operationId": "confirmar_adicional_v1_ordenes_trabajo__ot_id__confirmar_adicional_post",
        "parameters": [
          {
            "name": "ot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmarAdicionalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Confirmar Adicional V1 Ordenes Trabajo  Ot Id  Confirmar Adicional Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ordenes-trabajo/{ot_id}/autorizar-precio": {
      "post": {
        "tags": [
          "taller"
        ],
        "summary": "EP-TAL-05: Autorizar visibilidad de precio al cliente",
        "operationId": "autorizar_precio_v1_ordenes_trabajo__ot_id__autorizar_precio_post",
        "parameters": [
          {
            "name": "ot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutorizarPrecioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Autorizar Precio V1 Ordenes Trabajo  Ot Id  Autorizar Precio Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ordenes-trabajo/{ot_id}/revision-final": {
      "post": {
        "tags": [
          "taller"
        ],
        "summary": "EP-TAL-06: Declarar revisión final → REVISION_FINAL",
        "operationId": "revision_final_v1_ordenes_trabajo__ot_id__revision_final_post",
        "parameters": [
          {
            "name": "ot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionFinalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Revision Final V1 Ordenes Trabajo  Ot Id  Revision Final Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ordenes-trabajo/{ot_id}/cobro-parcial": {
      "post": {
        "tags": [
          "taller"
        ],
        "summary": "EP-TAL-07: Registrar cobro parcial con excepción 80%",
        "operationId": "cobro_parcial_v1_ordenes_trabajo__ot_id__cobro_parcial_post",
        "parameters": [
          {
            "name": "ot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CobroParcialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cobro Parcial V1 Ordenes Trabajo  Ot Id  Cobro Parcial Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ordenes-trabajo/{ot_id}/cerrar": {
      "post": {
        "tags": [
          "taller"
        ],
        "summary": "EP-TAL-08: Cerrar orden de trabajo",
        "operationId": "cerrar_ot_v1_ordenes_trabajo__ot_id__cerrar_post",
        "parameters": [
          {
            "name": "ot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cerrar Ot V1 Ordenes Trabajo  Ot Id  Cerrar Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ordenes-trabajo/{ot_id}/cancelar": {
      "post": {
        "tags": [
          "taller"
        ],
        "summary": "EP-TAL-09: Cancelar orden de trabajo",
        "operationId": "cancelar_ot_v1_ordenes_trabajo__ot_id__cancelar_post",
        "parameters": [
          {
            "name": "ot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelarOTRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancelar Ot V1 Ordenes Trabajo  Ot Id  Cancelar Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/ordenes-trabajo/{ot_id}": {
      "get": {
        "tags": [
          "taller"
        ],
        "summary": "EP-TAL-12: Obtener orden de trabajo",
        "operationId": "obtener_ot_v1_ordenes_trabajo__ot_id__get",
        "parameters": [
          {
            "name": "ot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Obtener Ot V1 Ordenes Trabajo  Ot Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbrirOTRequest": {
        "properties": {
          "vehiculo_id": {
            "type": "string",
            "title": "Vehiculo Id"
          },
          "mecanico_master_id": {
            "type": "string",
            "title": "Mecanico Master Id"
          },
          "modalidad": {
            "$ref": "#/components/schemas/ModalidadIntervencion"
          },
          "urgencia": {
            "$ref": "#/components/schemas/NivelUrgencia"
          },
          "mecanico_junior_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mecanico Junior Id"
          },
          "cliente_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cliente Id"
          }
        },
        "type": "object",
        "required": [
          "vehiculo_id",
          "mecanico_master_id",
          "modalidad",
          "urgencia"
        ],
        "title": "AbrirOTRequest"
      },
...
      "ConfirmarAdicionalRequest": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          }
        },
        "type": "object",
        "required": [
          "item_id"
        ],
        "title": "ConfirmarAdicionalRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}